% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Runlist.R
\name{add.blank}
\alias{add.blank}
\title{Insert blank rows}
\usage{
add.blank(Runlist, df.blank, blank.insert)
}
\arguments{
\item{Runlist}{A data frame to add blanks to.}

\item{df.blank}{A data frame consisting of only blanks to take from.}

\item{blank.insert}{An integer of how many blanks to add.}
}
\value{
Returns the same data frame that was supplied to the Runlist argument,
but with the specified number of appended blanks.
}
\description{
Function used within \code{\link[=create.Runlist]{create.Runlist()}} to add blanks.
}
\details{
Function used within \code{\link[=create.Runlist]{create.Runlist()}} to add blanks.
}
\examples{
library(dplyr)

full.list = Example_Runlist
Runlist = Runlist_final_empty
blank.max = 5
df.blank = dplyr::filter(full.list, full.list$LC_Well_Type == "blank") \%>\% # All blank rows
  dplyr::mutate(Draw_Max = blank.max, Draw_Count = 0)

Runlist = add.blank(Runlist, df.blank, 3) # Adds 3 blanks to Runlist.
Runlist
}
\seealso{
Other RunlistGenerator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.type}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{well.update}()}
}
\concept{RunlistGenerator}
