% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Runlist.R
\name{add.type}
\alias{add.type}
\title{Insert analyte rows for a well type}
\usage{
add.type(Runlist, df.analyte, compound, wellType)
}
\arguments{
\item{Runlist}{A data frame to add blanks to.}

\item{df.analyte}{A data frame with analytes to take from.}

\item{compound}{A string with what compound to filter by.}

\item{wellType}{A string with What well type to filter by.}
}
\value{
Returns the same data frame that was supplied to the Runlist argument,
but with appended analytes.
}
\description{
Function used within \code{\link[=create.Runlist]{create.Runlist()}} to add all analytes for a
specified compound and well type.
}
\examples{
library(dplyr)
full.list = Example_Runlist
df.analyte = dplyr::filter(full.list, full.list$LC_Well_Type == "Analyte") # All sample rows
Runlist <- add.type(Runlist, df.analyte, "Paracetamol", "cell")
Runlist
}
\seealso{
Other RunlistGenerator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.blank}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{well.update}()}
}
\concept{RunlistGenerator}
