% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Runlist.R
\name{create.Runlist}
\alias{create.Runlist}
\title{Create a Runlist from data frame}
\usage{
create.Runlist(
  full.list,
  blank.start = 3,
  blank.end = 5,
  blank.comp = 2,
  blank.type = 1,
  blank.max = 5
)
}
\arguments{
\item{full.list}{A full list, according to the package template, to be transformed
into a runlist.}

\item{blank.start}{Integer of blanks to start with.}

\item{blank.end}{Integer of blanks to end with.}

\item{blank.comp}{Integer of blanks to insert between compounds.}

\item{blank.type}{Integer of blanks to insert between well types.}

\item{blank.max}{Max number of times to draw from the same blank well.}
}
\value{
Returns a Runlist with blanks spaced as defined (with sensible defaults).
}
\description{
Creates a Runlist from the standardized data frame supplied by/made
in the Shiny app \code{\link[=RunlistGenerator]{RunlistGenerator()}}.
}
\details{
Creates a Runlist from the standardized data frame supplied by/made
in the Shiny app \code{\link[=RunlistGenerator]{RunlistGenerator()}}.
}
\examples{
head(Example_Runlist)
Runlist <- create.Runlist(full.list = Example_Runlist)
head(Runlist)
Runlist2 <- create.Runlist(full.list = Example_Runlist,
                          blank.start = 3,
                          blank.end = 5,
                          blank.comp = 2,
                          blank.type = 1,
                          blank.max = 5)
head(Runlist2)
}
\seealso{
Other Runlist Generator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.blank}()},
\code{\link{add.compound}()},
\code{\link{add.type}()},
\code{\link{aquaras.server}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{sanitizeInput}()},
\code{\link{well.update}()}
}
\concept{Runlist Generator}
