% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{dateSignAll.update}
\alias{dateSignAll.update}
\title{Update Date, Signature, & Sample_name}
\usage{
dateSignAll.update(df, date, signature)
}
\arguments{
\item{df}{Data frame to update}

\item{date}{New date value (character string)}

\item{signature}{New signature value}
}
\value{
Updated data frame
}
\description{
Update Date, Signature, & Sample_name for the entire data frame.
Sample_name is a composite in the format "date_signature_index".
}
\examples{
head(Example_Runlist)
Example_Runlist = dateSignAll.update(df = Example_Runlist,
                                     date = "20220730",
                                     signature = "RH")
head(Example_Runlist)
}
\seealso{
Other RunlistGenerator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.blank}()},
\code{\link{add.type}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{well.update}()}
}
\concept{RunlistGenerator}
