% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fufeces_runlist_template}
\alias{Fufeces_runlist_template}
\title{Fu feces runlist template}
\format{
A workbook object from openxlsx2
}
\source{
In house
}
\usage{
Fufeces_runlist_template
}
\description{
Workbook object from openxlsx2. Builds an .xlsx file acting
as a template for building a runlist that will be compatible
with \code{\link[=ras.Fu_feces_workflow]{ras.Fu_feces_workflow()}}.
}
\keyword{datasets}
