% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{loadFile.ML}
\alias{loadFile.ML}
\title{Load MassLynx complete summary file}
\usage{
loadFile.ML(sourceFile)
}
\arguments{
\item{sourceFile}{A file path}
}
\value{
A list of all lines in source file with "END" appended as the last line
}
\description{
Asks the user for a source file and loads it with \code{\link[readr:read_lines]{readr::read_lines()}}.
}
\examples{
\dontrun{
load.list = loadFile.ML(sourceFile)
}
}
\seealso{
Other SplitOutput: 
\code{\link{SplitOutput.ML}()},
\code{\link{cleanDF.ML}()},
\code{\link{splitDataLines.ML}()},
\code{\link{writeFiles.ML}()}
}
\concept{SplitOutput}
