% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.splitOutput.R
\name{ML.splitOutput}
\alias{ML.splitOutput}
\title{Split MassLynx output file}
\usage{
ML.splitOutput()
}
\value{
Writes a tsv file per compound to the same directory as the source file.
}
\description{
Splits the MassLynx complete summary output file into individual data frames
based on compound. These data frames are written to tsv files in the same
directory as the source file.
}
\examples{
ML.split.output() # First thing the function does is ask the user for a file.
}
