% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Runlist.R
\name{ras.add.compound}
\alias{ras.add.compound}
\title{Insert all analyte and blank rows for a compound}
\usage{
ras.add.compound(
  Runlist,
  df.analyte,
  df.blank,
  compound,
  wellType,
  blank.type = 1
)
}
\arguments{
\item{Runlist}{A data frame to add blanks to.}

\item{df.analyte}{A data frame with analytes to take from.}

\item{df.blank}{A data frame consisting of only blanks to take from.}

\item{compound}{A string with what compound to filter by.}

\item{wellType}{A vector with all well types to filter by. (string passed to
\code{\link[=ras.add.type]{ras.add.type()}})}

\item{blank.type}{An integer of how many blanks to add between types.
(default = 1; passed to \code{\link[=ras.add.blank]{ras.add.blank()}})}
}
\value{
Returns the same data frame that was supplied to the Runlist argument,
but appended with all analyte and blank rows for a compound.
}
\description{
Function used within \code{\link[=ras.create.Runlist]{ras.create.Runlist()}} to add all analyte and
blank rows for a compound. Passes arguments to \code{\link[=ras.add.type]{ras.add.type()}} and \code{\link[=ras.add.blank]{ras.add.blank()}}.
}
\examples{
# Setup
full.list = ras.Example_Runlist
blank.max = 5
df.blank = dplyr::filter(full.list, full.list$LC_Well_Type == "blank") \%>\% # All blank rows
  dplyr::mutate(Draw_Max = blank.max, Draw_Count = 0)
df.analyte = dplyr::filter(full.list, full.list$LC_Well_Type == "Analyte") # All sample rows
Runlist = tibble::tibble(
  Index        = double(),
  Plate        = double(),
  Row          = character(),
  Col          = double(),
  LC_Position  = character(),
  Date         = character(),
  Signature    = character(),
  Sample_name  = character(),
  Compound     = character(),
  Timepoint    = character(),
  Well_Type    = character(),
  LC_Well_Type = character(),
  Replicate    = character(),
  Sample_text  = character(),
  Draw_Max     = double(),
  Draw_Count   = double())


# Run function
  # Adds 1 blank between every type.
Runlist1 <- ras.add.compound(Runlist, df.analyte, df.blank, "Paracetamol",
  c("bead", "medium", "cell", "STD", "blank"))
Runlist1

  # Adds 3 blanks between every type.
Runlist2 <- ras.add.compound(Runlist, df.analyte, df.blank, "Ibuprofen",
  c("bead", "medium", "cell", "STD", "blank"), 3)
Runlist2
}
\seealso{
Other RunlistGenerator: 
\code{\link{ras.RunlistGenerator}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.type}()},
\code{\link{ras.create.Runlist}()},
\code{\link{ras.darkModeDF_Options}()},
\code{\link{ras.dateSignAll.update}()},
\code{\link{ras.sanitizeInput}()},
\code{\link{ras.server}()},
\code{\link{ras.well.update}()}
}
\concept{RunlistGenerator}
