% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Runlist.R
\name{ras.add.type}
\alias{ras.add.type}
\title{Insert analyte rows for a well type}
\usage{
ras.add.type(Runlist, df.analyte, compound, wellType)
}
\arguments{
\item{Runlist}{A data frame to add blanks to.}

\item{df.analyte}{A data frame with analytes to take from.}

\item{compound}{A string with what compound to filter by.}

\item{wellType}{A string with What well type to filter by.}
}
\value{
Returns the same data frame that was supplied to the Runlist argument,
but with appended analytes.
}
\description{
Function used within \code{\link[=ras.create.Runlist]{ras.create.Runlist()}} to add all analytes for a
specified compound and well type.
}
\examples{
# Setup
full.list = Example_Runlist
blank.max = 5
df.analyte = dplyr::filter(full.list, full.list$LC_Well_Type == "Analyte") # All sample rows
Runlist = tibble::tibble(
  Index        = double(),
  Plate        = double(),
  Row          = character(),
  Col          = double(),
  LC_Position  = character(),
  Date         = character(),
  Signature    = character(),
  Sample_name  = character(),
  Compound     = character(),
  Timepoint    = character(),
  Well_Type    = character(),
  LC_Well_Type = character(),
  Replicate    = character(),
  Sample_text  = character(),
  Draw_Max     = double(),
  Draw_Count   = double())

# Run function
Runlist1 <- ras.add.type(Runlist, df.analyte, "Paracetamol", "cell")
Runlist1
}
\seealso{
Other RunlistGenerator: 
\code{\link{aquaras.server}()},
\code{\link{ras.RunlistGenerator}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.compound}()},
\code{\link{ras.create.Runlist}()},
\code{\link{ras.darkModeDF_Options}()},
\code{\link{ras.dateSignAll.update}()},
\code{\link{ras.sanitizeInput}()},
\code{\link{ras.well.update}()}
}
\concept{RunlistGenerator}
