% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{ras.cleanDF}
\alias{ras.cleanDF}
\title{Clean list of data frames}
\usage{
ras.cleanDF(listDF)
}
\arguments{
\item{listDF}{List of data frames to clean}
}
\value{
List of data frames
}
\description{
Cleans each data frame by removing blanks and NAs. Also separates Name and
Sample Text into its composite columns.
}
\examples{
# Setup
dataLines = ras.loadFile(sourceFile = system.file("extdata",
                                                  "Example_MLOutput.txt",
                                                  package = "aquaras",
                                                  mustWork = TRUE))
### Ignore the "New names:" output.
listDF = ras.splitDataLines(dataLines)

listDF.clean = ras.cleanDF(listDF)
listDF.clean
}
\seealso{
Other SplitOutput: 
\code{\link{ras.SplitOutput}()},
\code{\link{ras.loadFile}()},
\code{\link{ras.splitDataLines}()},
\code{\link{ras.writeFiles}()}
}
\concept{SplitOutput}
