% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Runlist.R
\name{ras.create.Runlist}
\alias{ras.create.Runlist}
\title{Create a Runlist from data frame}
\usage{
ras.create.Runlist(
  full.list,
  blank.start = 3,
  blank.end = 5,
  blank.comp = 2,
  blank.type = 1,
  blank.max = 5
)
}
\arguments{
\item{full.list}{A full list, according to the package template, to be transformed
into a runlist.}

\item{blank.start}{Integer of blanks to start with.}

\item{blank.end}{Integer of blanks to end with.}

\item{blank.comp}{Integer of blanks to insert between compounds.}

\item{blank.type}{Integer of blanks to insert between well types.}

\item{blank.max}{Max number of times to draw from the same blank well.}
}
\value{
Returns a Runlist with blanks spaced as defined (with sensible defaults).
}
\description{
Creates a Runlist from the standardized data frame supplied by/made
in the Shiny app \code{\link[=ras.RunlistGenerator]{ras.RunlistGenerator()}}.
}
\examples{
# Setup
full.list = ras.Example_Runlist
head(full.list)

# Run function
Runlist1 <- ras.create.Runlist(full.list)
head(Runlist1)

Runlist2 <- ras.create.Runlist(full.list,
                               blank.start = 4,
                               blank.end   = 3,
                               blank.comp  = 2,
                               blank.type  = 1,
                               blank.max   = 6)
head(Runlist2)
}
\seealso{
Other RunlistGenerator: 
\code{\link{ras.RunlistGenerator}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.compound}()},
\code{\link{ras.add.type}()},
\code{\link{ras.darkModeDF_Options}()},
\code{\link{ras.dateSignAll.update}()},
\code{\link{ras.sanitizeInput}()},
\code{\link{ras.server}()},
\code{\link{ras.well.update}()}
}
\concept{RunlistGenerator}
