% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{ras.darkModeDF_Options}
\alias{ras.darkModeDF_Options}
\title{Options for data frame style}
\usage{
ras.darkModeDF_Options()
}
\description{
Sets \code{reactable.theme} in \code{\link[base:options]{base::options()}} with CSS styling borrowed from the \href{https://glin.github.io/reactable/articles/examples.html#theming}{Reactable example documentation}
}
\examples{
 \dontrun{
ras.darkModeDF_Options() # Sets the global options for reactable.theme
}
}
\seealso{
Other RunlistGenerator: 
\code{\link{ras.RunlistGenerator}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.compound}()},
\code{\link{ras.add.type}()},
\code{\link{ras.create.Runlist}()},
\code{\link{ras.dateSignAll.update}()},
\code{\link{ras.sanitizeInput}()},
\code{\link{ras.server}()},
\code{\link{ras.well.update}()}
}
\concept{RunlistGenerator}
