% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{ras.dateSignAll.update}
\alias{ras.dateSignAll.update}
\title{Update Date, Signature, & Sample_name}
\usage{
ras.dateSignAll.update(df, date, signature)
}
\arguments{
\item{df}{Data frame to update}

\item{date}{New date value (character string)}

\item{signature}{New signature value}
}
\value{
Updated data frame
}
\description{
Update Date, Signature, & Sample_name for the entire data frame.
Sample_name is a composite in the format "date_signature_index".
}
\examples{
head(Example_Runlist)
Example_Runlist1 = ras.dateSignAll.update(df = Example_Runlist,
                                          date      = "20220730",
                                          signature = "RH")
head(Example_Runlist1)
}
\seealso{
Other RunlistGenerator: 
\code{\link{aquaras.server}()},
\code{\link{ras.RunlistGenerator}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.compound}()},
\code{\link{ras.add.type}()},
\code{\link{ras.create.Runlist}()},
\code{\link{ras.darkModeDF_Options}()},
\code{\link{ras.sanitizeInput}()},
\code{\link{ras.well.update}()}
}
\concept{RunlistGenerator}
