% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fic_data_prosessing.R
\name{ras.Fic_timepoint}
\alias{ras.Fic_timepoint}
\title{Extract values for Cell & Medium (values w/ timepoints)}
\usage{
ras.Fic_timepoint(
  df,
  values = "Conc.",
  types = c("Cell", "Medium"),
  .summarize = TRUE
)
}
\arguments{
\item{df}{Data frame from \code{ras.Fic_cleanup()}.}

\item{values}{Name of the column to use for values.}

\item{types}{Vector with the types to select timepoints for. (Cell & Medium)}

\item{.summarize}{Summarize values into averages.}
}
\value{
A list of data frames with the selected timepoints.
}
\description{
Extract values for Cell & Medium (values w/ timepoints)
}
\examples{
if (interactive()) {
  df_ex <- data.frame(
    Sample_ID = rep(c("A", "B"), each = 4),
    Sample_type = rep(c("Cell", "Cell", "Medium", "Medium"), times = 2),
    Timepoint = rep(c(30, 60, 30, 60), times = 2),
    `Conc.` = c(10, 15, 9, 11, 20, 24, 17, 19)
  )

  ras.Fic_timepoint(df_ex)
}
}
