% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fic_workflow.R
\name{ras.Fic_workflow}
\alias{ras.Fic_workflow}
\title{Fic end-to-end workflow (!!!)}
\usage{
ras.Fic_workflow(
  source = c("Waters", "Sciex"),
  ID = "Sample_ID",
  values = "Conc.",
  Buffer = "HBSS",
  Dilution_type = "[:digit:]x",
  Dilution_extract = "[:digit:]+(?=x)",
  stab = "Stab",
  czero = "Czero",
  prot_cell_value = "mg_Protein",
  prot_cell_type = "Cells",
  prot_hom_value = "Protein_conc._mg/mL",
  prot_hom_type = "hom",
  V.medium = 200,
  MassBalance_2.5 = 1.75,
  .compound = "Analyte Peak Name",
  .checkValues = TRUE,
  prot_.split = "Sample Text",
  .summarize = TRUE,
  .SD = TRUE
)
}
\arguments{
\item{source}{Data source (\code{"Sciex"} or \code{"Waters"}).}

\item{ID}{Column name for ID column after splitting.}

\item{values}{Column name for numeric values.}

\item{Buffer}{Regex used to select buffer values.}

\item{Dilution_type}{Regex used to select dilution-factor values.}

\item{Dilution_extract}{Regex used to extract the dilution factor.}

\item{stab}{Regex used to select stability values.}

\item{czero}{Regex used to select C-zero values.}

\item{prot_cell_value}{Column name for values}

\item{prot_cell_type}{RegEx to filter by}

\item{prot_hom_value}{RegEx to filter by}

\item{prot_hom_type}{RegEx to filter by}

\item{V.medium}{RegEx to filter by}

\item{MassBalance_2.5}{Factor for calculating Mass Balance 10.2.5.}

\item{.compound}{Column name for compounds to prefix Sample_ID,
\code{NULL} => "Analyte Peak Name"}

\item{.checkValues}{Removes "<" from the values column to make sure it can
be coerced to numeric.}

\item{prot_.split}{Column name to split for samples in protein dataframe}

\item{.summarize}{Summarize the data into unique rows in
grouping column. Drops the original value columns.}

\item{.SD}{Create column with Standard Deviation}
}
\value{
Data frame with all variables used and calculated.
}
\description{
Currently not functional!
}
\examples{
if (interactive()) {
  ras.Fic_workflow(source = "Sciex")
}
}
