% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fic_calculations.R
\name{ras.Fu_feces_meanSD}
\alias{ras.Fu_feces_meanSD}
\title{Fu feces mean & SD}
\usage{
ras.Fu_feces_meanSD(
  df,
  grouping_col = "Sample_ID",
  sample_type = "Sample Type",
  compound = "Analyte Peak Name"
)
}
\arguments{
\item{df}{Data frame}

\item{grouping_col}{Column name to group by}

\item{sample_type}{Column name for sample type}

\item{compound}{Column name for compound}
}
\value{
Data frame with mean & SD columns
}
\description{
Add mean & SD columns, grouped by \code{grouping_col}, for all
the calculated numeric columns.
}
\examples{
df_ex <- data.frame(
  Sample_ID = c("S1", "S1", "S2", "S2"),
  `Sample Type` = c("hom", "hom", "hom", "hom"),
  `Analyte Peak Name` = c("cmpd", "cmpd", "cmpd", "cmpd"),
  value = c(10, 12, 20, 22)
)

ras.Fu_feces_meanSD(df_ex)
}
