% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fu_feces_workflow.R
\name{ras.Fu_feces_workflow}
\alias{ras.Fu_feces_workflow}
\title{Fu feces workflow wrapper}
\usage{
ras.Fu_feces_workflow(
  source = c("Sciex", "Waters"),
  ID = "Sample Name",
  Sample_type = "Sample Type",
  values = "Calculated Concentration (nM)",
  Buffer = "HBSS",
  Dilution_type = "[:digit:]+x",
  Dilution_extract = "[:digit:]+(?=x$)",
  Dilution_factor = 1,
  stab = "Stab",
  czero = "Czero",
  D = (1/((1/4.8) * 0.95)),
  MassBalance_2.5 = 1.75,
  .compound = "Analyte Peak Name",
  .checkValues = TRUE,
  .summarize = FALSE
)
}
\arguments{
\item{source}{Data source (\code{"Sciex"} or \code{"Waters"}).}

\item{ID}{Column name for ID column after splitting.}

\item{Sample_type}{Name of the type column in raw data.}

\item{values}{Column name for numeric values.}

\item{Buffer}{Regex used to select buffer values.}

\item{Dilution_type}{Regex used to select dilution-factor values.}

\item{Dilution_extract}{Regex used to extract the dilution factor.}

\item{Dilution_factor}{Numeric value for dilution factor}

\item{stab}{Regex used to select stability values.}

\item{czero}{Regex used to select C-zero values.}

\item{D}{Numeric value for D.}

\item{MassBalance_2.5}{Factor for calculating Mass Balance 10.2.5.}

\item{.compound}{Column name for compounds to prefix Sample_ID,
\code{NULL} => "Analyte Peak Name"}

\item{.checkValues}{Removes "<" from the values column to make sure it can
be coerced to numeric.}

\item{.summarize}{Summarize the data into unique rows in
grouping column. Drops the original value columns.}
}
\value{
Data frame with all variables used and calculated.
Also writes the data frame to \code{.csv} (EU) and \code{.xlsx} files.
}
\description{
Fu feces workflow wrapper
}
\examples{
if (interactive()) {
  ras.Fu_feces_workflow(source = "Sciex")
}
}
