% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{ras.loadFile}
\alias{ras.loadFile}
\title{Load MassLynx complete summary file}
\usage{
ras.loadFile(sourceFile)
}
\arguments{
\item{sourceFile}{A file path}
}
\value{
A list of all lines in source file with "END" appended as the last line
}
\description{
Asks the user for a source file and loads it with \code{\link[readr:read_lines]{readr::read_lines()}}.
}
\examples{
  \dontrun{
dataLines = ras.loadFile(sourceFile = system.file("extdata",
                                                  "Example_MLOutput.txt",
                                                  package = "aquaras",
                                                  mustWork = TRUE))
head(dataLines)
}
}
\seealso{
Other SplitOutput: 
\code{\link{ras.SplitOutput}()},
\code{\link{ras.cleanDF}()},
\code{\link{ras.splitDataLines}()},
\code{\link{ras.writeFiles}()}
}
\concept{SplitOutput}
