% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGroups.R
\name{ras.ran_group}
\alias{ras.ran_group}
\title{Random group}
\usage{
ras.ran_group(
  masterList = master_list,
  groupName = c(),
  seriesList = series_list,
  seriesCheck = series_check,
  groupSize = group_size
)
}
\arguments{
\item{masterList}{A character vector with observations/compound names}

\item{groupName}{Empty vector to fill with observations}

\item{seriesList}{Data frame used for control of how many times an observation occurs}

\item{seriesCheck}{Character vector that has the cumulative unique observations across the series}

\item{groupSize}{Integer of how many observations to attempt to place in each group}
}
\value{
List with 1. populated \code{groupName}, 2. updated \code{seriesList}, and 3. updated \code{seriesCheck}
}
\description{
Populate a group with random observations/compounds. Used inside \code{\link[=ras.ran_series]{ras.ran_series()}}.
}
