% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGroups.R
\name{ras.ran_series}
\alias{ras.ran_series}
\title{Random series of groups}
\usage{
ras.ran_series(master_list, group_size)
}
\arguments{
\item{master_list}{A character vector with observations/compound names}

\item{group_size}{Integer of how many observations to attempt to place in each group}
}
\value{
A list of character vectors with each random group (data frame?)
}
\description{
Creates a series of groups with randomly sampled observations. Observations
are distributed such as that there are no repeated entries between groups/within
a series. Number of groups in the series is calculated as \code{length(master_list)}
divided by \code{group_size}, rounded up to nearest integer. If an observation can
not be found (for example because the groups can not be filled evenly) an error
message is printed to the console and \code{N/A} inserted instead.
}
\examples{
# Example with 4 even groups
# Compound list
Compound_list = c(
"Comp_1", "Comp_2", "Comp_3", "Comp_4", "Comp_5",
"Comp_6", "Comp_7", "Comp_8", "Comp_9", "Comp_10",
"Comp_11", "Comp_12", "Comp_13", "Comp_14", "Comp_15",
"Comp_16", "Comp_17", "Comp_18", "Comp_19", "Comp_20",
"Comp_21", "Comp_22", "Comp_23", "Comp_24")

series_A = ras.ran_series(master_list = Compound_list, group_size = 6)
series_A

# Example with 5 uneven groups
# Compound list
Compound_list = c(
"Comp_1", "Comp_2", "Comp_3", "Comp_4", "Comp_5",
"Comp_6", "Comp_7", "Comp_8", "Comp_9", "Comp_10",
"Comp_11", "Comp_12", "Comp_13", "Comp_14", "Comp_15",
"Comp_16", "Comp_17", "Comp_18", "Comp_19", "Comp_20",
"Comp_21", "Comp_22", "Comp_23", "Comp_24", "Comp_25")

series_B = ras.ran_series(Compound_list, group_size = 6)
series_B

}
