% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGroups.R
\name{ras.Randomizer}
\alias{ras.Randomizer}
\title{Randomizer}
\usage{
ras.Randomizer(
  compound_list,
  group_size,
  series = 1,
  write = FALSE,
  fileName = NULL
)
}
\arguments{
\item{compound_list}{A character vector with observations/compound names}

\item{group_size}{Integer of how many observations to attempt to place in each group}

\item{series}{Number of series to make}

\item{write}{Default \code{FALSE}. If set to \code{TRUE} will write series to a .xlsx file}

\item{fileName}{Default \code{NULL}. Name of file. Automatically appended with \code{.xlsx}}
}
\value{
List of data frames
}
\description{
Uses \code{\link[=ras.ran_series]{ras.ran_series()}} to create a series of groups with randomly sampled observations.
Observations are distributed such as that there are no repeated entries between groups/within
a series. Number of groups in the series is calculated as \code{length(master_list)}
divided by \code{group_size}, rounded up to nearest integer. If an observation can
not be found (for example because the groups can not be filled evenly) an error
message is printed to the console and \code{N/A} inserted instead.
If \code{write = TRUE} will write the series to a comma delimited .csv file (excel?).
}
