% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ras.RunlistGenerator}
\alias{ras.RunlistGenerator}
\title{App for making a runlist}
\usage{
ras.RunlistGenerator()
}
\value{
Creates a Shiny GUI.
}
\description{
Program to input info for 96-well-plates. Generates a runlist for the Waters LC/MS software MassLynx.
}
\examples{
 \dontrun{
ras.RunlistGenerator() # Launches the Shiny GUI.
}
}
\seealso{
Other RunlistGenerator: 
\code{\link{aquaras.server}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.compound}()},
\code{\link{ras.add.type}()},
\code{\link{ras.create.Runlist}()},
\code{\link{ras.darkModeDF_Options}()},
\code{\link{ras.dateSignAll.update}()},
\code{\link{ras.sanitizeInput}()},
\code{\link{ras.well.update}()}
}
\concept{RunlistGenerator}
