% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{ras.sanitizeInput}
\alias{ras.sanitizeInput}
\title{Sanitize input}
\usage{
ras.sanitizeInput(inputVar)
}
\arguments{
\item{inputVar}{Character string to sanitize}
}
\value{
Character string
}
\description{
Replaces underscores with periods.
}
\examples{
ras.sanitizeInput("Spaces etc -- but_no_underscore!")
}
\seealso{
Other RunlistGenerator: 
\code{\link{ras.RunlistGenerator}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.compound}()},
\code{\link{ras.add.type}()},
\code{\link{ras.create.Runlist}()},
\code{\link{ras.darkModeDF_Options}()},
\code{\link{ras.dateSignAll.update}()},
\code{\link{ras.server}()},
\code{\link{ras.well.update}()}
}
\concept{RunlistGenerator}
