% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{ras.server}
\alias{ras.server}
\title{Server for \code{\link[=ras.RunlistGenerator]{ras.RunlistGenerator()}}}
\usage{
ras.server(input, output, session)
}
\arguments{
\item{input, output, session}{As described under \code{server} in \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\description{
Fulfills the \code{server} argument in \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.
}
\examples{
 \dontrun{
shiny::shinyApp(ui = ras.ui, server = ras.server)
}
}
\seealso{
Other RunlistGenerator: 
\code{\link{ras.RunlistGenerator}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.compound}()},
\code{\link{ras.add.type}()},
\code{\link{ras.create.Runlist}()},
\code{\link{ras.darkModeDF_Options}()},
\code{\link{ras.dateSignAll.update}()},
\code{\link{ras.sanitizeInput}()},
\code{\link{ras.well.update}()}
}
\concept{RunlistGenerator}
