% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_setup.R
\name{ras.setup_dir_and_subfolders}
\alias{ras.setup_dir_and_subfolders}
\title{Setup directory & subfolders}
\usage{
ras.setup_dir_and_subfolders(output_dir = NULL)
}
\arguments{
\item{output_dir}{A directory path}
}
\value{
A directory path
}
\description{
Setup directory & subfolders
}
\examples{
out_dir <- file.path(tempdir(), "aquaras_example_output")
subdirs <- ras.setup_dir_and_subfolders(out_dir)
names(subdirs)
}
