% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{ras.splitDataLines}
\alias{ras.splitDataLines}
\title{Split loaded file into data frames}
\usage{
ras.splitDataLines(dataLines)
}
\arguments{
\item{dataLines}{List of vectors from \code{\link[readr:read_lines]{readr::read_lines()}} in \code{\link[=ras.loadFile]{ras.loadFile()}}}
}
\value{
List of data frames
}
\description{
Split the file loaded with \code{\link[=ras.loadFile]{ras.loadFile()}} into dataframes by compound.
}
\examples{
  \dontrun{
# Setup
dataLines = ras.loadFile(sourceFile = system.file("extdata",
                                                  "Example_MLOutput.txt",
                                                  package = "aquaras",
                                                  mustWork = TRUE))
### Ignore the "New names:" output.
listDF = ras.splitDataLines(dataLines)
listDF
}

}
\seealso{
Other SplitOutput: 
\code{\link{ras.SplitOutput}()},
\code{\link{ras.cleanDF}()},
\code{\link{ras.loadFile}()},
\code{\link{ras.writeFiles}()}
}
\concept{SplitOutput}
