% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{ras.StackOutput}
\alias{ras.StackOutput}
\title{Stack MassLynx output file}
\usage{
ras.StackOutput(
  sourceFiles = tcltk::tk_choose.files(caption = "Select MassLynx output file"),
  write = TRUE
)
}
\arguments{
\item{sourceFiles}{A character vector with one or more file paths}

\item{write}{Defaults to TRUE for writing to file system}
}
\value{
Writes a tsv file per compound to the same directory as the source file.

Also returns the nested list of data frames for each file
}
\description{
Stacks the MassLynx complete summary output file into a single dataframe.
Adds a column with what compound was used.
Behaves like \code{\link[=ras.SplitOutput]{ras.SplitOutput()}}.
\subsection{write = TRUE}{

\strong{!!! DEFAULT IS TO WRITE TO FILE SYSTEM !!!} Data frames written to tsv files
in the same directory as the source file. Use write = FALSE to disable this.
}
}
\examples{
  \dontrun{
  # See ras.SplitOutput() for now
  }
}
\seealso{
Other SplitOutput: 
\code{\link{ras.SplitOutput}()}
}
\concept{SplitOutput}
