% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_avg_StDev}
\alias{ras.TPA_avg_StDev}
\title{TPA - Average & StDev/Range}
\usage{
ras.TPA_avg_StDev(df, sample_names)
}
\arguments{
\item{df}{Data frame to add columns to}

\item{sample_names}{Data frame from \code{ras.TPA_sample_names}}
}
\value{
Same data frame with new columns
}
\description{
Adds columns for average and standard deviation (or range if there are only two sample cols).
}
\examples{
### Very specific function; no example yet
}
