% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_avg_StDev_calc}
\alias{ras.TPA_avg_StDev_calc}
\title{TPA - Average & StDev/Range}
\usage{
ras.TPA_avg_StDev_calc(df, sample_names, na.rm = FALSE)
}
\arguments{
\item{df}{Data frame to add columns to}

\item{sample_names}{List of samples from \code{ras.TPA_sample_names}}

\item{na.rm}{If TRUE, NA values are dropped before determining calculation}
}
\value{
Same data frame with new columns
}
\description{
Adds columns for average and standard deviation (or range if there are only two sample cols).
}
\examples{
### Very specific function; no example yet
}
