% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_barplot_helper}
\alias{ras.TPA_barplot_helper}
\title{TPA - Barplot helper function}
\usage{
ras.TPA_barplot_helper(row_v, helper_args)
}
\arguments{
\item{row_v}{Named character vector row from apply}

\item{cur_sam}{Current sample}

\item{gene_col_RegEx}{RegEx for selecting the gene name column/value}

\item{save}{If \code{TRUE}, saves plot to directory}

\item{extension}{What extension to use for both file ending and \code{\link[=ggsave]{ggsave()}} \code{device} argument}

\item{directory}{Directory path to write to}
}
\value{
One ggplot object
}
\description{
Helper function that does the actual plotting and saving.
}
