% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_barplot_save}
\alias{ras.TPA_barplot_save}
\title{TPA - Barplot & save}
\usage{
ras.TPA_barplot_save(
  df,
  gene_col_RegEx = "[Gg]ene",
  sample_names = NULL,
  save = FALSE,
  extension = "png",
  directory = NA
)
}
\arguments{
\item{df}{Data frame reshaped by \code{ras.TPA_reshape_filter}}

\item{save}{While \code{TRUE} writes files to system}

\item{extension}{File extension/type for exported images and \code{\link[=ggsave]{ggsave()}} \code{device} argument}

\item{directory}{Path to destination folder}

\item{genes}{Character vector of what genes to plot}
}
\value{
Nested list of all plots generated
}
\description{
Create barplots and (if save = TRUE) save them to file system.
}
\examples{
### Very specific function; no example yet
}
