% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_calcFromIntensity}
\alias{ras.TPA_calcFromIntensity}
\title{TPA - Calculation}
\usage{
ras.TPA_calcFromIntensity(
  df_raw,
  metaData_RegEx =
    "protein_names|gene_names|majority_protein|razor_unique_peptides$|reverse|potential_contaminant|mol_weight",
  dropIntensity = TRUE
)
}
\arguments{
\item{df_raw}{Data frame (from MaxQuant)}

\item{metaData_RegEx}{Character string for regular expression identifying
columns to keep (passed to \code{\link[base:grep]{base::grep()}}). Must include "mol_weight"!}

\item{dropIntensity}{If to drop intensity (default = TRUE)}
}
\value{
Data frame with meta data and TPA values
}
\description{
Calculate TPA from intensity using the raw output file from MaxQuant.
The function cleans the column names and keeps only columns identified by the
\code{metaData_RegEx}pattern and containing "intensity". TPA is calculated for
each column while preserving the experiment identifier. If \code{dropIntensity}
= FALSE intensity columns are preserved.
}
\examples{
  \dontrun{
# my_TPA <- ras.TPA_calcFromIntensity(df_raw = my_data)

# my_TPA2 <- ras.TPA_calcFromIntensity(df_raw = my_data,
#               metaData_RegEx = "protein_names|mol_weight")
}

}
