% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_reshape_filter}
\alias{ras.TPA_reshape_filter}
\title{TPA – Reshape & filter}
\usage{
ras.TPA_reshape_filter(df, gene_regex = "*")
}
\arguments{
\item{df}{Data frame from \code{ras.TPA_avg_StDev}}

\item{gene_regex}{Regular expression for what genes to filter for}
}
\value{
Data frame with sample values in long format
}
\description{
Pivots the data frame into a longer format creating the columns \code{sample} and \code{StDev_and_Range}
}
\examples{
### Very specific function; no example yet
}
