% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_sample_names}
\alias{ras.TPA_sample_names}
\title{TPA - Sample names}
\usage{
ras.TPA_sample_names(df)
}
\arguments{
\item{df}{Data frame with samples in columns}
}
\value{
List with two data frames, (sample_names) & (sample_groups). Each with two columns, (1) unique samples & (2) how many replicates.
}
\description{
Extracts unique sample names and counts how many repeats it has.
}
\examples{
### Very specific function; no example yet
}
