% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_StDev_range_DeterminationForRows}
\alias{ras.TPA_StDev_range_DeterminationForRows}
\title{TPA - Determine and calculate StDev or Range}
\usage{
ras.TPA_StDev_range_DeterminationForRows(df, na.rm = FALSE)
}
\arguments{
\item{df}{Data frame passed from \code{\link[=ras.TPA_avg_StDev_calc]{ras.TPA_avg_StDev_calc()}}}

\item{na.rm}{If TRUE, drops NA values before determining calculation}
}
\value{
A vector to be inserted as a column in the original data frame
}
\description{
Internal function.
}
\details{
Checks number of values in a row and calculates SD, range, or inputs NA.

If \code{na.rm = TRUE} it drops NA values from the row first (doesn't affect the final data frame).
}
