% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{ras.well.update}
\alias{ras.well.update}
\title{Update well info}
\usage{
ras.well.update(
  df,
  well.current,
  date,
  signature,
  compound,
  timepoint,
  type,
  replicate
)
}
\arguments{
\item{df}{Data frame to update}

\item{well.current}{Current well plate position (LC_Position)}

\item{date}{New date value (character string)}

\item{signature}{New signature value}

\item{compound}{New compound value}

\item{timepoint}{New timepoint value}

\item{type}{New type value}

\item{replicate}{New replicate value}
}
\value{
Same data frame, but with updated row
}
\description{
Update well info
}
\examples{
Example_Runlist[1,]
Example_Runlist1 = ras.well.update(df = Example_Runlist,
                                   well.current = "3:A,1",
                                   date         = "20220730",
                                   signature    = "RH",
                                   compound     = "Warfarin",
                                   timepoint    = "42",
                                   type         = "bead",
                                   replicate    = "6")
Example_Runlist1[1,]
}
\seealso{
Other RunlistGenerator: 
\code{\link{aquaras.server}()},
\code{\link{ras.RunlistGenerator}()},
\code{\link{ras.add.blank}()},
\code{\link{ras.add.compound}()},
\code{\link{ras.add.type}()},
\code{\link{ras.create.Runlist}()},
\code{\link{ras.darkModeDF_Options}()},
\code{\link{ras.dateSignAll.update}()},
\code{\link{ras.sanitizeInput}()}
}
\concept{RunlistGenerator}
