% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{ras.writeFiles}
\alias{ras.writeFiles}
\title{Write data frames to files}
\usage{
ras.writeFiles(listDF, sourceFile)
}
\arguments{
\item{listDF}{List of data frames to write to files}

\item{sourceFile}{Source file for locating what directory to write to}
}
\description{
Writes data frames to individual .txt files (tsv) in the same directory as source file.
}
\examples{
  \dontrun{
ras.writeFiles(listDF, sourceFile)
}
}
\seealso{
Other SplitOutput: 
\code{\link{ras.SplitOutput}()},
\code{\link{ras.cleanDF}()},
\code{\link{ras.loadFile}()},
\code{\link{ras.splitDataLines}()}
}
\concept{SplitOutput}
