% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{RunlistGenerator}
\alias{RunlistGenerator}
\title{App for making a runlist}
\usage{
RunlistGenerator()
}
\value{
Creates a Shiny GUI.
}
\description{
Program to input info for 96-well-plates. Generates a runlist for the Waters LC/MS software MassLynx.
}
\examples{
RunlistGenerator() # Launches the Shiny GUI.
}
