% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{sanitizeInput}
\alias{sanitizeInput}
\title{Sanitize input}
\usage{
sanitizeInput(inputVar)
}
\arguments{
\item{inputVar}{Character string to sanitize}
}
\value{
Character string
}
\description{
Replaces spaces and underscores with periods.
}
