% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.splitOutput.R
\name{splitDataLines.ML}
\alias{splitDataLines.ML}
\title{Split loaded file into data frames}
\usage{
splitDataLines.ML(dataLines)
}
\arguments{
\item{dataLines}{List of vectors from readr::read_lines() in loadFile.ML()}
}
\value{
List of data frames
}
\description{
Split the file loaded with loadFile.ML() into dataframes by compound.
}
\examples{
listDF = splitDataLines.ML(dataLines)
}
