% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Runlist.R
\name{add.blank}
\alias{add.blank}
\title{Insert blank rows}
\usage{
add.blank(Runlist, df.blank, blank.insert)
}
\arguments{
\item{Runlist}{A data frame to add blanks to.}

\item{df.blank}{A data frame consisting of only blanks to take from.}

\item{blank.insert}{An integer of how many blanks to add.}
}
\value{
Returns the same data frame that was supplied to the Runlist argument,
but with the specified number of appended blanks.
}
\description{
Function used within \code{\link[=create.Runlist]{create.Runlist()}} to add blanks.
}
\details{
Function used within \code{\link[=create.Runlist]{create.Runlist()}} to add blanks.
}
\examples{
# Setup
full.list = Example_Runlist
blank.max = 5
df.blank = dplyr::filter(full.list, full.list$LC_Well_Type == "blank") \%>\% # All blank rows
  dplyr::mutate(Draw_Max = blank.max, Draw_Count = 0)
Runlist = tibble::tibble(
  Index        = double(),
  Plate        = double(),
  Row          = character(),
  Col          = double(),
  LC_Position  = character(),
  Date         = character(),
  Signature    = character(),
  Sample_name  = character(),
  Compound     = character(),
  Timepoint    = character(),
  Well_Type    = character(),
  LC_Well_Type = character(),
  Replicate    = character(),
  Sample_text  = character(),
  Draw_Max     = double(),
  Draw_Count   = double())

# Run function
Runlist1 = add.blank(Runlist, df.blank, 3) # Adds 3 blanks to Runlist.
Runlist1
}
\seealso{
Other RunlistGenerator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.compound}()},
\code{\link{add.type}()},
\code{\link{aquaras.server}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{sanitizeInput}()},
\code{\link{well.update}()}
}
\concept{RunlistGenerator}
