% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Runlist.R
\name{add.blank}
\alias{add.blank}
\title{Insert blank rows}
\usage{
add.blank(Runlist, df.blank, blank.insert)
}
\arguments{
\item{Runlist}{A data frame to add blanks to.}

\item{df.blank}{A data frame consisting of only blanks to take from.}

\item{blank.insert}{An integer of how many blanks to add.}
}
\value{
Returns the same data frame that was supplied to the Runlist argument,
but with the specified number of appended blanks.
}
\description{
Function used within \code{\link[=create.Runlist]{create.Runlist()}} to add blanks.
}
\details{
Function used within \code{\link[=create.Runlist]{create.Runlist()}} to add blanks.
}
\examples{
\dontrun{ Runlist <- add.blank(Runlist, df.blank, 3) # Adds 3 blanks (from df.blank) to Runlist. }
}
\seealso{
Other Runlist Generator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.compound}()},
\code{\link{add.type}()},
\code{\link{aquaras.server}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{sanitizeInput}()},
\code{\link{well.update}()}
}
\concept{Runlist Generator}
