% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquaras-package.R
\docType{package}
\name{aquaras-package}
\alias{aquaras}
\alias{aquaras-package}
\title{aquaras: Data processing related to MassLynx}
\description{
Created for personal use. Provides tools for working with runlists to use with MassLynx (Waters LC/MS software). Calling RunlistGenerator() launches a Shiny app that lets you create the full data frame and a runlist. Alternatively, create.Runlist() may be used directly.
}
\keyword{internal}
