% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{aquaras.server}
\alias{aquaras.server}
\title{Server for \code{\link[=RunlistGenerator]{RunlistGenerator()}}}
\usage{
aquaras.server(input, output, session)
}
\description{
Server for \code{\link[=RunlistGenerator]{RunlistGenerator()}}
}
