% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{aquaras.server}
\alias{aquaras.server}
\title{Server for \code{\link[=RunlistGenerator]{RunlistGenerator()}}}
\usage{
aquaras.server(input, output, session)
}
\arguments{
\item{input, output, session}{As described under \code{server} in \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\description{
Fulfills the \code{server} argument in \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.
}
\examples{
\dontrun{
shiny::shinyApp(ui = ui, server = aquaras.server)
}
}
\seealso{
Other Runlist Generator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.blank}()},
\code{\link{add.compound}()},
\code{\link{add.type}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{sanitizeInput}()},
\code{\link{well.update}()}
}
\concept{Runlist Generator}
