% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.splitOutput.R
\name{cleanDF.ML}
\alias{cleanDF.ML}
\title{Clean list of data frames}
\usage{
cleanDF.ML(listDF)
}
\arguments{
\item{listDF}{List of data frames to clean}
}
\value{
List of data frames
}
\description{
Cleans each data frame by removing blanks and NAs. Also separates Name and
Sample Text into its composite columns.
}
\examples{
listDF = cleanDF.ML(listDF)
}
