% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{darkModeDF_Options}
\alias{darkModeDF_Options}
\title{Options for data frame style}
\usage{
darkModeDF_Options()
}
\description{
Sets \code{reactable.theme} in \code{\link[base:options]{base::options()}} with CSS styling borrowed from the \href{https://glin.github.io/reactable/articles/examples.html#theming}{Reactable example documentation}
}
\examples{
\dontrun{
darkModeDF_Options() # Sets the global options for reactable.theme
}
}
\seealso{
Other RunlistGenerator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.blank}()},
\code{\link{add.type}()},
\code{\link{create.Runlist}()},
\code{\link{dateSignAll.update}()},
\code{\link{well.update}()}
}
\concept{RunlistGenerator}
