% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{dateSignAll.update}
\alias{dateSignAll.update}
\title{Update Date, Signature, & Sample_name}
\usage{
dateSignAll.update(df, date, signature)
}
\arguments{
\item{df}{Data frame to use}

\item{date}{New date (character string)}

\item{signature}{New signature (character string)}
}
\value{
Updated data frame
}
\description{
Update Date, Signature, & Sample_name for the entire data frame.
Sample_name is a composite in the format "date_signature_index".
}
\examples{
Runlist = update.dateSignAll(Runlist, "20220722", "RH")
}
