% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fic.R
\name{ras.Fic_dataImport}
\alias{ras.Fic_dataImport}
\title{Load data for Fic & Fu feces}
\usage{
ras.Fic_dataImport(source, .compound)
}
\arguments{
\item{source}{A string, either \code{"Sciex"} or \code{"Waters"}}

\item{.compound}{.compound as supplied to the workflow function}
}
\value{
A list with the loaded data frame & appropriate
value for .compound
}
\description{
Load data for Fic & Fu feces
}
\examples{
  # No example
}
