% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fic.R
\name{ras.Fic_timepoint}
\alias{ras.Fic_timepoint}
\title{Extract values for Cell & Medium (values w/ timepoints)}
\usage{
ras.Fic_timepoint(
  df,
  values = "Conc.",
  types = c("Cell", "Medium"),
  .summarize = TRUE
)
}
\arguments{
\item{df}{Dataframe from ras.Fic_cleanup().}

\item{values}{Name of the column to use for values.}

\item{types}{Vector with the types to select timepoints for. (Cell & Medium)}
}
\value{
List of dataframes with the selected timepoints
}
\description{
Extract values for Cell & Medium (values w/ timepoints)
}
\examples{
 \dontrun{
 # No example yet
 }
}
