% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fic.R
\name{ras.Fic_workflow}
\alias{ras.Fic_workflow}
\title{Fic end-to-end workflow}
\usage{
ras.Fic_workflow(
  source = c("Waters", "Sciex"),
  ID = "Sample_ID",
  values = "Conc.",
  Buffer = "HBSS",
  Dilution_type = "[:digit:]x",
  Dilution_extract = "[:digit:]+(?=x)",
  stab = "Stab$",
  czero = "Czero$",
  Kp = "Kp$",
  prot_cell_value = "mg_Protein",
  prot_cell_type = "Cells",
  prot_hom_value = "Protein_conc._mg/mL",
  prot_hom_type = "hom",
  V.medium = 200
)
}
\arguments{
\item{source}{What source the data file comes from}

\item{ID}{Column name for ID column after splitting}

\item{values}{Column name for values in data}

\item{Buffer}{RegEx to filter by for buffer values}

\item{Dilution_type}{RegEx to filter by for dilution factor values}

\item{Dilution_extract}{RegEx for extracting the dilution factor}

\item{stab}{RegEx to filter by for stability values}

\item{czero}{RegEx to filter by for C zero values values}

\item{Kp}{RegEx to filter by for kp values}

\item{prot_cell_value}{Column name for values}

\item{prot_cell_type}{RegEx to filter by}

\item{prot_hom_value}{RegEx to filter by}

\item{prot_hom_type}{RegEx to filter by}

\item{V.medium}{RegEx to filter by}
}
\value{
Dataframe with all variables, used & calculated
}
\description{
Fic end-to-end workflow
}
\examples{
  \dontrun{
  # No example yet
  }
}
