% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fic_workflow.R
\name{ras.Fu_feces_workflow}
\alias{ras.Fu_feces_workflow}
\title{Fu feces workflow wrapper}
\usage{
ras.Fu_feces_workflow(
  source = c("Sciex", "Waters"),
  ID = "Sample Name",
  Sample_type = "Sample Type",
  values = "Calculated Concetration (nM)",
  Buffer = "HBSS",
  Dilution_type = "[:digit:]x",
  Dilution_extract = "[:digit:]+(?=x)",
  Dilution_factor = 1,
  stab = "Stab",
  czero = "Czero",
  D = 4.8,
  .compound = "Analyte Peak Name",
  .checkValues = TRUE,
  .summarize = TRUE,
  .SD = TRUE
)
}
\arguments{
\item{source}{What source the data file comes from}

\item{ID}{Column name for ID column after splitting}

\item{Sample_type}{Name for type column in raw data}

\item{values}{Column name for values in data}

\item{Buffer}{RegEx to filter by for buffer values}

\item{Dilution_type}{RegEx to filter by for dilution factor values}

\item{Dilution_extract}{RegEx for extracting the dilution factor}

\item{Dilution_factor}{Numeric value for dilution factor}

\item{stab}{RegEx to filter by for stability values}

\item{czero}{RegEx to filter by for C zero values values}

\item{D}{Numeric value for D}

\item{.compound}{Column name for compounds to prefix Sample_ID,
\code{NULL} => "Analyte Peak Name"}

\item{.checkValues}{Removes "<" from the values column to make sure it can
be coerced to numeric.}

\item{.summarize}{Summarize values into averages}

\item{.SD}{Create column with Standard Deviation}
}
\value{
Dataframe with all variables, used & calculated
}
\description{
Fu feces workflow wrapper
}
\examples{
   \dontrun{
  # No example yet
  }
}
