% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{ras.SplitOutput}
\alias{ras.SplitOutput}
\title{Split MassLynx output file}
\usage{
ras.SplitOutput(
  sourceFiles = tcltk::tk_choose.files(),
  clean = FALSE,
  write = TRUE
)
}
\arguments{
\item{sourceFiles}{A character vector with one or more file paths}

\item{clean}{Defaults to FALSE for data cleaning}

\item{write}{Defaults to TRUE for writing to file system}
}
\value{
Writes a tsv file per compound to the same directory as the source file.

Also returns the nested list of data frames for each file
}
\description{
Splits the MassLynx complete summary output file into individual data frames
based on compound.
\subsection{Uninteresting output}{

Spliting the file generates a stream of \code{"New names: • `` -> `...1`"}
output in the console that is not particularly interesting but lets you know
it's doing something.
}

\subsection{clean = TRUE}{

These data frames are cleaned by ras.cleanDF() (unless clean = FALSE).
}

\subsection{write = TRUE}{

\strong{!!! DEFAULT IS TO WRITE TO FILE SYSTEM !!!} Data frames written to tsv files
in the same directory as the source file. Use write = FALSE to disable this.
}
}
\examples{
  \dontrun{
ras.SplitOutput() # First thing the function does is ask the user for a file.
ras.SplitOutput(clean = FALSE) # If the summary output file was not based on the provided template.


# !!! DOES **NOT** WRITE TO FILE SYSTEM !!!
listDF = ras.SplitOutput(sourceFile = system.file("extdata",
                                                  "Example_MLOutput.txt",
                                                  package = "aquaras",
                                                  mustWork = TRUE),
                         write = FALSE)
listDF
}

}
\seealso{
Other SplitOutput: 
\code{\link{ras.StackOutput}()}
}
\concept{SplitOutput}
