% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_barplot_save}
\alias{ras.TPA_barplot_save}
\title{TPA – Barplot & save}
\usage{
ras.TPA_barplot_save(
  df,
  genes,
  extension = "png",
  directory = tcltk::tk_choose.dir(),
  save = TRUE
)
}
\arguments{
\item{df}{Data frame reshaped by \code{ras.TPA_reshape_filter}}

\item{genes}{Character vector of what genes to plot}

\item{extension}{File extension/type for exported images}

\item{directory}{Path to destination folder}

\item{save}{While \code{TRUE} (default) writes files to system}
}
\value{
\code{NULL}
}
\description{
Create barplots and (if save = TRUE) save them to file system.
}
\examples{
### Very specific function; no example yet
}
