% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPA_sample_names}
\alias{ras.TPA_sample_names}
\title{TPA - Sample names}
\usage{
ras.TPA_sample_names(df)
}
\arguments{
\item{df}{Data frame with samples in columns}
}
\value{
Data frame with three columns, (1) unique samples, (2) how many replicates, and (3) the sample group
}
\description{
Extracts unique sample names and counts how many repeats it has.
}
\examples{
### Very specific function; no example yet
}
