% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPAer}
\alias{ras.TPAer}
\title{TPA - Do all the TPA stuff}
\usage{
ras.TPAer(
  df,
  na.rm = FALSE,
  save = FALSE,
  file_type = "png",
  folder_path = tcltk::tk_choose.dir()
)
}
\arguments{
\item{df}{Data frame with values}

\item{na.rm}{If \code{TRUE}, NA values are dropped before determining calculation}

\item{save}{If \code{TRUE}, writes files to system}

\item{file_type}{File extension/type for exported images}

\item{folder_path}{Path to destination folder}
}
\value{
List containing (1) Data frame and (2) List of generated plots
}
\description{
Clean, restructure, calculate, plot, and export/save plots to file system.
}
\examples{
### Very specific function; no example yet
}
