% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPA.R
\name{ras.TPAer}
\alias{ras.TPAer}
\title{TPA - Do all the TPA stuff}
\usage{
ras.TPAer(
  df,
  gene_regex = "*",
  file_type = "png",
  folder_path = tcltk::tk_choose.dir(),
  save = TRUE
)
}
\arguments{
\item{df}{Data frame with values}

\item{gene_regex}{Regular expression for which genes to plot}

\item{file_type}{File extension/type for exported images}

\item{folder_path}{Path to destination folder}

\item{save}{While \code{TRUE} (default) writes files to system}
}
\value{
List containing (1) Data frame and (2) List of generated plots
}
\description{
Clean, restructure, pivot longer, subset, plot, and export/save plots to file system.
}
\examples{
### Very specific function; no example yet
}
