% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{RunlistGenerator}
\alias{RunlistGenerator}
\title{App for making a runlist}
\usage{
RunlistGenerator()
}
\value{
Creates a Shiny GUI.
}
\description{
Program to input info for 96-well-plates. Generates a runlist for the Waters LC/MS software MassLynx.
}
\examples{
RunlistGenerator() # Launches the Shiny GUI.
}
\seealso{
Other Runlist Generator: 
\code{\link{add.blank}()},
\code{\link{add.compound}()},
\code{\link{add.type}()},
\code{\link{aquaras.server}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{sanitizeInput}()},
\code{\link{well.update}()}
}
\concept{Runlist Generator}
