% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{sanitizeInput}
\alias{sanitizeInput}
\title{Sanitize input}
\usage{
sanitizeInput(inputVar)
}
\arguments{
\item{inputVar}{Character string to sanitize}
}
\value{
Character string
}
\description{
Replaces underscores with periods.
}
\examples{
sanitizeInput("Spaces etc -- but_no_underscore!")
}
\seealso{
Other Runlist Generator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.blank}()},
\code{\link{add.compound}()},
\code{\link{add.type}()},
\code{\link{aquaras.server}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{well.update}()}
}
\concept{Runlist Generator}
