% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{splitDataLines.ML}
\alias{splitDataLines.ML}
\title{Split loaded file into data frames}
\usage{
splitDataLines.ML(dataLines)
}
\arguments{
\item{dataLines}{List of vectors from \code{\link[readr:read_lines]{readr::read_lines()}} in \code{\link[=loadFile.ML]{loadFile.ML()}}}
}
\value{
List of data frames
}
\description{
Split the file loaded with \code{\link[=loadFile.ML]{loadFile.ML()}} into dataframes by compound.
}
\examples{
listDF = splitDataLines.ML(dataLines)
}
