% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitOutput.ML.R
\name{SplitOutput.ML}
\alias{SplitOutput.ML}
\title{Split MassLynx output file}
\usage{
<<<<<<< HEAD
SplitOutput.ML(clean = TRUE)
=======
splitOutput.ML(clean = TRUE)
>>>>>>> 14f428ad5efa263519fdfbf42b9b3cc4d157080d
}
\arguments{
\item{clean}{Defaults to TRUE,}
}
\value{
Writes a tsv file per compound to the same directory as the source file.
}
\description{
Splits the MassLynx complete summary output file into individual data frames
based on compound. These data frames are cleaned by \code{\link[=cleanDF.ML]{cleanDF.ML()}} (unless
clean = FALSE) and written to tsv files in the same directory as the source file.
}
\examples{
<<<<<<< HEAD
SplitOutput.ML() # First thing the function does is ask the user for a file.
SplitOutput.ML(clean = FALSE) # If the summary output file was not based on the provided template.
=======
splitOutput.ML() # First thing the function does is ask the user for a file.
splitOutput.ML(clean = FALSE) # If the summary output file was not based on the provided template.
>>>>>>> 14f428ad5efa263519fdfbf42b9b3cc4d157080d
}
\seealso{
Other Split Output: 
\code{\link{cleanDF.ML}()},
\code{\link{loadFile.ML}()},
\code{\link{splitDataLines.ML}()},
\code{\link{writeFiles.ML}()}
}
\concept{Split Output}
