% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{well.update}
\alias{well.update}
\title{Update well info}
\usage{
well.update(
  df,
  well.current,
  date,
  signature,
  compound,
  timepoint,
  type,
  replicate
)
}
\arguments{
\item{df}{Data frame to update row}

\item{well.current}{Current well plate position (LC_Position)}

\item{date}{New date value (character string)}

\item{signature}{New signature value}

\item{compound}{New compound value}

\item{timepoint}{New timepoint value}

\item{type}{New type value}

\item{replicate}{New replicate value}
}
\value{
Same data frame, but with updated row
}
\description{
Update well info
}
\seealso{
Other Runlist Generator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.blank}()},
\code{\link{add.compound}()},
\code{\link{add.type}()},
\code{\link{aquaras.server}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()},
\code{\link{sanitizeInput}()}
}
\concept{Runlist Generator}
