% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunc_server.R
\name{well.update}
\alias{well.update}
\title{Update well info}
\usage{
well.update(
  df,
  well.current,
  date,
  signature,
  compound,
  timepoint,
  type,
  replicate
)
}
\arguments{
\item{df}{Data frame to update}

\item{well.current}{Current well plate position (LC_Position)}

\item{date}{New date value (character string)}

\item{signature}{New signature value}

\item{compound}{New compound value}

\item{timepoint}{New timepoint value}

\item{type}{New type value}

\item{replicate}{New replicate value}
}
\value{
Same data frame, but with updated row
}
\description{
Update well info
}
\examples{
Example_Runlist[1,]
Example_Runlist = well.update(df = Example_Runlist,
                              well.current = "3:A,1",
                              date = "20220730",
                              signature = "RH",
                              compound = "Warfarin",
                              timepoint = "42",
                              type = "bead",
                              replicate = "6")
Example_Runlist[1,]
}
\seealso{
Other RunlistGenerator: 
\code{\link{RunlistGenerator}()},
\code{\link{add.blank}()},
\code{\link{add.type}()},
\code{\link{create.Runlist}()},
\code{\link{darkModeDF_Options}()},
\code{\link{dateSignAll.update}()}
}
\concept{RunlistGenerator}
