% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOutput.ML.R
\name{writeFiles.ML}
\alias{writeFiles.ML}
\title{Write data frames to files}
\usage{
writeFiles.ML(listDF, sourceFile)
}
\arguments{
\item{listDF}{List of data frames to write to files}

\item{sourceFile}{Source file for locating what directory to write to}
}
\description{
Writes data frames to individual .txt files (tsv) in the same directory as source file.
}
\examples{
  \dontrun{
writeFiles.ML(listDF, sourceFile)
}
}
\seealso{
Other SplitOutput: 
\code{\link{SplitOutput.ML}()},
\code{\link{cleanDF.ML}()},
\code{\link{loadFile.ML}()},
\code{\link{splitDataLines.ML}()}
}
\concept{SplitOutput}
